backend/requirements.txt
torch
torchvision
scikit-learn>=1.6
joblib
numpy<2
Pillow
fastapi
uvicorn[standard]
python-multipart
